issue 16: change itsYourTurn to return the move
